[#330581 #447119] 4. Разделен контролер объявлений на два: работа сосписок объявлений, работа с конкретным объявлением
