+++ src/test/java/ru/job4j/cars/controller/PostControllerTest.java
@@ -10,7 +10,6 @@ import org.mockito.MockitoAnnotations;
 import org.springframework.ui.Model;
 import org.springframework.web.multipart.MultipartFile;
 import ru.job4j.cars.common.dto.DetailsResponse;
-import ru.job4j.cars.common.dto.PostListResponse;
 import ru.job4j.cars.common.model.car.Car;
 import ru.job4j.cars.common.model.user.User;
 import ru.job4j.cars.common.security.UserSession;
@@ -24,7 +23,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.math.BigDecimal;
-import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -76,71 +74,6 @@ class PostControllerTest {
     }
 
     @Test
-    public void whenGetPostList() {
-        List<PostListResponse> posts = List.of(new PostListResponse());
-        when(postService.getAllIsNotSold()).thenReturn(posts);
-
-        String actual = controller.getPostList(model, httpSession);
-
-        assertNotNull(actual);
-        assertEquals("post/post", actual);
-        verify(postService).getAllIsNotSold();
-        verifyNoMoreInteractions(postService);
-    }
-
-    @Test
-    public void whenPostListForDay() {
-        List<PostListResponse> posts = List.of(new PostListResponse());
-        when(postService.getAllForDay()).thenReturn(posts);
-
-        String actual = controller.getPostListForDay(model, httpSession);
-
-        assertNotNull(actual);
-        assertEquals("post/post", actual);
-        verify(postService).getAllForDay();
-        verifyNoMoreInteractions(postService);
-    }
-
-    @Test
-    public void whenGetPostListWithMileage() {
-        List<PostListResponse> posts = List.of(new PostListResponse());
-        when(postService.findAllPostByCategoryName("С пробегом")).thenReturn(posts);
-
-        String actual = controller.getPostListWithMileage(model, httpSession);
-
-        assertNotNull(actual);
-        assertEquals("post/post", actual);
-        verify(postService).findAllPostByCategoryName("С пробегом");
-        verifyNoMoreInteractions(postService);
-    }
-
-    @Test
-    public void whenGetPostListNew() {
-        List<PostListResponse> posts = List.of(new PostListResponse());
-        when(postService.findAllPostByCategoryName("Новые")).thenReturn(posts);
-
-        String actual = controller.getPostListNew(model, httpSession);
-
-        assertNotNull(actual);
-        assertEquals("post/post", actual);
-        verify(postService).findAllPostByCategoryName("Новые");
-        verifyNoMoreInteractions(postService);
-    }
-
-    @Test
-    public void whenGetPostListAll() {
-        List<PostListResponse> posts = List.of(new PostListResponse());
-        when(postService.getAll()).thenReturn(posts);
-
-        String actual = controller.getPostListAll(model, httpSession);
-
-        assertNotNull(actual);
-        assertEquals("post/post", actual);
-        verify(postService).getAll();
-        verifyNoMoreInteractions(postService);
-    }
-
-    @Test
     public void whenGetDetails() {
         DetailsResponse details = new DetailsResponse();
         when(postService.getById(ID, ID)).thenReturn(details);
@@ -180,7 +113,7 @@ class PostControllerTest {
         String actual = controller.create(user, new Car(), "", 0, file, 0);
 
         assertNotNull(actual);
-        assertEquals("redirect:/post/my", actual);
+        assertEquals("redirect:/posts/my", actual);
         verify(postService).create(any());
         verify(file).getOriginalFilename();
         verify(file).getBytes();
@@ -234,7 +167,7 @@ class PostControllerTest {
         String actual = controller.delete(model, httpSession, ID);
 
         assertNotNull(actual);
-        assertEquals("redirect:/post/my/", actual);
+        assertEquals("redirect:/posts/my/", actual);
         verify(postService).delete(ID);
         verifyNoMoreInteractions(postService);
     }
